Detangle base-services from instrumentation-api
