+++ platforms/core-runtime/base-services/build.gradle.kts
@@ -34,7 +34,6 @@ dependencies {
     api(libs.jsr305)
     api(libs.guava)
 
-    implementation(projects.internalInstrumentationApi)
     implementation(projects.io)
     implementation(projects.time)
