+++ platforms/core-runtime/base-services/src/main/java/org/gradle/internal/work/DefaultWorkerLimits.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,18 @@
  * limitations under the License.
  */
 
-/**
- * Classes related to Gradle parallelism and concurrency.
- */
-package org.gradle.concurrent;
+package org.gradle.internal.work;
+
+public class DefaultWorkerLimits implements WorkerLimits {
+
+    private final int maxWorkerCount;
+
+    public DefaultWorkerLimits(int maxWorkerCount) {
+        this.maxWorkerCount = maxWorkerCount;
+    }
+
+    @Override
+    public int getMaxWorkerCount() {
+        return maxWorkerCount;
+    }
+}
