+++ platforms/core-runtime/base-services/src/main/java/org/gradle/internal/work/WorkerLimits.java
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.concurrent;
+package org.gradle.internal.work;
 
-import org.gradle.api.NonNullApi;
 import org.gradle.internal.service.scopes.Scope;
 import org.gradle.internal.service.scopes.ServiceScope;
 
@@ -32,7 +31,6 @@ import org.gradle.internal.service.scopes.ServiceScope;
  *     <li>A language compiler in a forked process</li>
  * </ul>
  */
-@NonNullApi
 @ServiceScope(Scope.CrossBuildSession.class)
 public interface WorkerLimits {
