+++ subprojects/core-api/src/main/java/org/gradle/concurrent/ParallelismConfiguration.java
