+++ subprojects/core/src/test/groovy/org/gradle/execution/plan/DefaultPlanExecutorTest.groovy
@@ -23,9 +23,9 @@ import org.gradle.api.internal.tasks.TaskStateInternal
 import org.gradle.api.invocation.Gradle
 import org.gradle.initialization.BuildCancellationToken
 import org.gradle.internal.buildoption.DefaultInternalOptions
-import org.gradle.internal.concurrent.DefaultWorkerLimits
 import org.gradle.internal.concurrent.ExecutorFactory
 import org.gradle.internal.resources.DefaultResourceLockCoordinationService
+import org.gradle.internal.work.DefaultWorkerLimits
 import org.gradle.internal.work.WorkerLeaseRegistry
 import org.gradle.internal.work.WorkerLeaseService
 import spock.lang.Specification
