+++ subprojects/core/src/test/groovy/org/gradle/internal/model/StateTransitionControllerTest.groovy
@@ -19,9 +19,9 @@ package org.gradle.internal.model
 import org.gradle.internal.Describables
 import org.gradle.internal.Factory
 import org.gradle.internal.build.ExecutionResult
-import org.gradle.internal.concurrent.DefaultWorkerLimits
 import org.gradle.internal.resources.DefaultResourceLockCoordinationService
 import org.gradle.internal.work.DefaultWorkerLeaseService
+import org.gradle.internal.work.DefaultWorkerLimits
 import org.gradle.test.fixtures.concurrent.ConcurrentSpec
 
 import java.util.function.Function
