+++ subprojects/core/src/test/groovy/org/gradle/internal/operations/MaxWorkersTest.groovy
@@ -16,9 +16,9 @@
 
 package org.gradle.internal.operations
 
-import org.gradle.internal.concurrent.DefaultWorkerLimits
 import org.gradle.internal.resources.DefaultResourceLockCoordinationService
 import org.gradle.internal.work.DefaultWorkerLeaseService
+import org.gradle.internal.work.DefaultWorkerLimits
 import org.gradle.internal.work.WorkerLeaseService
 import org.gradle.test.fixtures.concurrent.ConcurrentSpec
